# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  email: string;
  name: string;
  picture?: string;
  isOwner: boolean;
  role: string;
}

export interface InventoryItem {
  location: string;
  vehicle: string;
  vin: string;
  price: string;
  km?: string;
  carfax?: string;
  website?: string;
  onlinePrice?: string;
  matrixPrice?: string | null;
  cost?: string | null;
  bbAvgWholesale?: string | null;
}

export interface CacheStatus {
  lastUpdated?: string | null;
  isRefreshing: boolean;
  count: number;
}

export interface VehicleImages {
  vin: string;
  urls: string[];
}

export interface AccessEntry {
  email: string;
  addedAt: string;
  addedBy: string;
  role: string;
}

export interface AddAccessRequest {
  email: string;
  role?: string;
}

export interface UpdateAccessRoleRequest {
  role: string;
}

export interface AuditLogEntry {
  id: number;
  action: string;
  targetEmail: string;
  changedBy: string;
  roleFrom?: string | null;
  roleTo?: string | null;
  timestamp: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  ok: boolean;
}

export interface LenderProgramTier {
  tierName: string;
  maxAdvanceLTV: number;
  maxAftermarketLTV: number;
  maxAllInLTV: number;
  creditorFee: number;
  dealerReserve: number;
  minRate?: number | null;
  maxRate?: number | null;
  minTerm?: number | null;
  maxTerm?: number | null;
}

export interface LenderProgram {
  lenderCode: string;
  lenderName: string;
  tiers: LenderProgramTier[];
}

export interface LenderProgramsResponse {
  programs: LenderProgram[];
  updatedAt?: string | null;
  sourceApp?: string | null;
}

export interface LenderStatus {
  running: boolean;
  startedAt?: string | null;
  lastRun?: string | null;
  lenderCount: number;
  error?: string | null;
  programsAge?: string | null;
}

export interface LenderCalculateRequest {
  lenderCode: string;
  tierName: string;
  approvedRate: number;
  approvedTerm: number;
  maxPayment: number;
  downPayment?: number;
  tradeValue?: number;
  tradeLien?: number;
  taxRate?: number;
  includeAftermarket?: boolean;
  aftermarketAmount?: number;
}

export interface LenderCalcResultItem {
  vin: string;
  vehicle: string;
  location: string;
  bbWholesale: number;
  maxAdvance: number;
  totalFinanced: number;
  monthlyPayment: number;
  costOfBorrowing: number;
  ltv: number;
  hasPhotos?: boolean;
  website?: string;
}

export interface LenderCalculateResponse {
  lender: string;
  tier: string;
  tierConfig: LenderProgramTier;
  resultCount: number;
  results: LenderCalcResultItem[];
}

export type GetVehicleImagesParams = {
  vin: string;
};