# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current authenticated user
 */
export const GetMeResponse = zod.object({
  email: zod.string(),
  name: zod.string(),
  picture: zod.string().optional(),
  isOwner: zod.boolean(),
  role: zod.string(),
});

/**
 * @summary Get all inventory items
 */
export const GetInventoryResponseItem = zod.object({
  location: zod.string(),
  vehicle: zod.string(),
  vin: zod.string(),
  price: zod.string(),
  km: zod.string().optional(),
  carfax: zod.string().optional(),
  website: zod.string().optional(),
  onlinePrice: zod.string().optional(),
  matrixPrice: zod.string().nullish(),
  cost: zod.string().nullish(),
  bbAvgWholesale: zod.string().nullish(),
});
export const GetInventoryResponse = zod.array(GetInventoryResponseItem);

/**
 * @summary Get the timestamp of the last inventory cache refresh
 */
export const GetCacheStatusResponse = zod.object({
  lastUpdated: zod.string().nullish(),
  isRefreshing: zod.boolean(),
  count: zod.number(),
});

/**
 * @summary Get photo gallery URLs for a vehicle by VIN
 */
export const GetVehicleImagesQueryParams = zod.object({
  vin: zod.coerce.string(),
});

export const GetVehicleImagesResponse = zod.object({
  vin: zod.string(),
  urls: zod.array(zod.string()),
});

/**
 * @summary Get list of approved emails (owner only)
 */
export const GetAccessListResponseItem = zod.object({
  email: zod.string(),
  addedAt: zod.string(),
  addedBy: zod.string(),
  role: zod.string(),
});
export const GetAccessListResponse = zod.array(GetAccessListResponseItem);

/**
 * @summary Add an email to the access list (owner only)
 */
export const AddAccessEntryBody = zod.object({
  email: zod.string(),
  role: zod.string().optional(),
});

export const AddAccessEntryResponse = zod.object({
  email: zod.string(),
  addedAt: zod.string(),
  addedBy: zod.string(),
  role: zod.string(),
});

/**
 * @summary Update a user's role (owner only)
 */
export const UpdateAccessRoleParams = zod.object({
  email: zod.coerce.string(),
});

export const UpdateAccessRoleBody = zod.object({
  role: zod.string(),
});

export const UpdateAccessRoleResponse = zod.object({
  email: zod.string(),
  addedAt: zod.string(),
  addedBy: zod.string(),
  role: zod.string(),
});

/**
 * @summary Remove an email from the access list (owner only)
 */
export const RemoveAccessEntryParams = zod.object({
  email: zod.coerce.string(),
});

export const RemoveAccessEntryResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * @summary Get cached lender program matrices (owner only)
 */
export const GetLenderProgramsResponse = zod.object({
  programs: zod.array(
    zod.object({
      lenderCode: zod.string(),
      lenderName: zod.string(),
      tiers: zod.array(
        zod.object({
          tierName: zod.string(),
          maxAdvanceLTV: zod.number(),
          maxAftermarketLTV: zod.number(),
          maxAllInLTV: zod.number(),
          creditorFee: zod.number(),
          dealerReserve: zod.number(),
          minRate: zod.number().nullish(),
          maxRate: zod.number().nullish(),
          minTerm: zod.number().nullish(),
          maxTerm: zod.number().nullish(),
        }),
      ),
    }),
  ),
  updatedAt: zod.string().nullish(),
  sourceApp: zod.string().nullish(),
});

/**
 * @summary Get lender sync status (owner only)
 */
export const GetLenderStatusResponse = zod.object({
  running: zod.boolean(),
  startedAt: zod.string().nullish(),
  lastRun: zod.string().nullish(),
  lenderCount: zod.number(),
  error: zod.string().nullish(),
  programsAge: zod.string().nullish(),
});

/**
 * @summary Trigger manual lender sync (owner only)
 */
export const RefreshLenderResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * @summary Calculate inventory affordability by lender/tier (owner only)
 */
export const LenderCalculateBody = zod.object({
  lenderCode: zod.string(),
  tierName: zod.string(),
  approvedRate: zod.number(),
  approvedTerm: zod.number(),
  maxPayment: zod.number(),
  downPayment: zod.number().optional(),
  tradeValue: zod.number().optional(),
  tradeLien: zod.number().optional(),
  taxRate: zod.number().optional(),
  includeAftermarket: zod.boolean().optional(),
  aftermarketAmount: zod.number().optional(),
});

export const LenderCalculateResponse = zod.object({
  lender: zod.string(),
  tier: zod.string(),
  tierConfig: zod.object({
    tierName: zod.string(),
    maxAdvanceLTV: zod.number(),
    maxAftermarketLTV: zod.number(),
    maxAllInLTV: zod.number(),
    creditorFee: zod.number(),
    dealerReserve: zod.number(),
    minRate: zod.number().nullish(),
    maxRate: zod.number().nullish(),
    minTerm: zod.number().nullish(),
    maxTerm: zod.number().nullish(),
  }),
  resultCount: zod.number(),
  results: zod.array(
    zod.object({
      vin: zod.string(),
      vehicle: zod.string(),
      location: zod.string(),
      bbWholesale: zod.number(),
      maxAdvance: zod.number(),
      totalFinanced: zod.number(),
      monthlyPayment: zod.number(),
      costOfBorrowing: zod.number(),
      ltv: zod.number(),
      hasPhotos: zod.boolean().optional(),
      website: zod.string().optional(),
    }),
  ),
});

/**
 * @summary Get audit log of access changes (owner only)
 */
export const GetAuditLogResponseItem = zod.object({
  id: zod.number(),
  action: zod.string(),
  targetEmail: zod.string(),
  changedBy: zod.string(),
  roleFrom: zod.string().nullish(),
  roleTo: zod.string().nullish(),
  timestamp: zod.string(),
});
export const GetAuditLogResponse = zod.array(GetAuditLogResponseItem);