# lib/api-zod/src/generated/types/lenderCalcResultItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LenderCalcResultItem {
  vin: string;
  vehicle: string;
  location: string;
  bbWholesale: number;
  maxAdvance: number;
  totalFinanced: number;
  monthlyPayment: number;
  costOfBorrowing: number;
  ltv: number;
  hasPhotos?: boolean;
  website?: string;
}