# lib/api-zod/src/generated/types/lenderCalculateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LenderCalculateRequest {
  lenderCode: string;
  tierName: string;
  approvedRate: number;
  approvedTerm: number;
  maxPayment: number;
  downPayment?: number;
  tradeValue?: number;
  tradeLien?: number;
  taxRate?: number;
  includeAftermarket?: boolean;
  aftermarketAmount?: number;
}