# lib/api-zod/src/generated/types/lenderCalculateResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LenderCalcResultItem } from "./lenderCalcResultItem";
import type { LenderProgramTier } from "./lenderProgramTier";

export interface LenderCalculateResponse {
  lender: string;
  tier: string;
  tierConfig: LenderProgramTier;
  resultCount: number;
  results: LenderCalcResultItem[];
}