# lib/api-zod/src/generated/types/lenderProgram.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LenderProgramTier } from "./lenderProgramTier";

export interface LenderProgram {
  lenderCode: string;
  lenderName: string;
  tiers: LenderProgramTier[];
}