# lib/api-zod/src/generated/types/lenderProgramTier.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LenderProgramTier {
  tierName: string;
  maxAdvanceLTV: number;
  maxAftermarketLTV: number;
  maxAllInLTV: number;
  creditorFee: number;
  dealerReserve: number;
  minRate?: number | null;
  maxRate?: number | null;
  minTerm?: number | null;
  maxTerm?: number | null;
}