# lib/api-zod/src/generated/types/lenderProgramsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LenderProgram } from "./lenderProgram";

export interface LenderProgramsResponse {
  programs: LenderProgram[];
  updatedAt?: string | null;
  sourceApp?: string | null;
}