# lib/api-zod/src/generated/types/lenderStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LenderStatus {
  running: boolean;
  startedAt?: string | null;
  lastRun?: string | null;
  lenderCount: number;
  error?: string | null;
  programsAge?: string | null;
}